+++ src/powerquery-language-services/externalType/externalTypeUtils.ts
@@ -5,6 +5,22 @@ import { Type } from "@microsoft/powerquery-parser/lib/powerquery-parser/languag
 
 import { ExternalType } from ".";
 
+export function composeExternalTypeResolvers(
+    ...resolverFns: ReadonlyArray<ExternalType.TExternalTypeResolverFn>
+): ExternalType.TExternalTypeResolverFn {
+    return (request: ExternalType.TExternalTypeRequest): Type.TPowerQueryType | undefined => {
+        for (const resolverFn of resolverFns) {
+            const resolvedType: Type.TPowerQueryType | undefined = resolverFn(request);
+
+            if (resolvedType !== undefined) {
+                return resolvedType;
+            }
+        }
+
+        return undefined;
+    };
+}
+
 export function createValueTypeRequest(identifierLiteral: string): ExternalType.ExternalValueTypeRequest {
     return {
         kind: ExternalType.ExternalTypeRequestKind.Value,
